From: Keith Bostic <bostic@vangogh.CS.Berkeley.edu>
Here are the patches that make stty support vmin/vtime set
and display.

I'm not sure that reset should also handle this problem.  (I
wouldn't mind if reset were to fix this, but I'd argue that it
should do a full "stty sane" then, not just add vmin/vtime to
what it already does.)

--keith