Adjust the name of the included file to match reality.
Somehow, I don't think this stuff was tested at all! :-(
I really hope that it actually works, though my hopes are steadily diminishing.
Anyone with 27xx/28xx boards in -current is *strongly encouraged* to give this
stuff a shot!  Otherwise, I suspect that we'll be punting this out of
2.0.  I haven't found a single part of Justin's commit that wasn't broken
in some way.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@4603 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f