route.c: keep track of where cloned routes come from, and make sure to delete them when the ``parent'' goes away
route.h: add glue to track this to rtentry structure.  WARNING WILL ROBINSON!
This will be yet another incompatible change in your route-using binaries.
I apologize, but this was the only way to do it.  I took this opportunity
to increase the size of the metrics to what I believe will be the final
length for 2.1, so that when the T/TCP stuff is done, this won't happen
again.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@5791 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f