Finally applying my own fix. :-)
>Description:

ctm(1) sometimes did not free up all used resources (open pipes and
processes, heap memory).  This happened whenever one of the passes
ended prematurely, and it became very apparent when running it on
a bunch of already applied deltas, resulting in a ``gunzip: resource
temporarily unavailable'' due to the maxproc # exhausted.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@7168 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f