Argh, I hate the tcl build/config environment. :-(
It seems that some tools (eg: expect-5.21) use TCL_LIB_SPEC to generate
their shared library suffix.  This should be .so.1.0, not .so as ld can't
use it.

Revert part of the previous change here, it did too much.  libtcl75.so.1.1
was bumped, but the rules to generate library names for _other_ packages
were not supposed to be. Sigh.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@19263 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f