I haven't had a single contraversial commit all week, so what the heck.  Watch through our hidden camera, ladies and gentlemen, as this one-line addition to the syslog output generates hundreds of thousands of lines of email in response, all from people decrying the evils of electronic noise pollution! :-)
What this change does, simply speaking, is syslog it every time
someone changes their local password.  I need this at a local ISP to
tell whether people are reacting to expires in a timely fashion or
not.  To disable it, uncomment -DLOGGING in the Makefile.

If your users change their passwords so often as to fill your logfile,
then you may also have another administrative problem to deal with.