Move _JBLEN definition to <machine/setjmp.h> like on NetBSD. Remove definitions for architectures that we don't have libc MD source for.
Change 'int' to 'long' in sigjmp_buf and jmp_buf structures so
that they are suitable for the register size on both i386 and alpha.
Although FreeBSD defines these structures differently to NetBSD,
the element size is now the same.

If Bruce doesn't like this [8-)], we'll need to add a typedef to
each <machine/setjmp.h> and use that in <setjmp.h>.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@32419 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f