Taking the GENERIC kernel and commenting out MSDOSFS, CD9660{,_ROOT}, PROCFS, NFS_ROOT will produce kernel that cannot mount a UFS /.
Vfs type numbers must be distinct from VFS_GENERIC (and VFS_VFSCONF, but
that has the same value and should go away).

The problem happens because NFS is the first vfs (in sys/conf order) so it
gets type number 0 and conflicts harmfully with VFS_GENERIC which is also 0.
The conflict is apparently harmless in the usual case when another vfs
gets type number 0, because nfs is the only vfs that has sysctls.

Inital fix by:	Dima <dima@tejblum.dnttm.rssi.ru>
Reason why it worked by:	bde


git-svn-id: svn+ssh://svn.freebsd.org/base/head@39975 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f