Add TARGET_ARCH=mipsel and TARGET_ARCH=mipseb as a valid targets. Unlike the unisex architecutres we've had so far, mips is bisexual. These tools can produce either byte sex, and the compiler/make determines the proper gender to use.  Otherwise, we'd have to have had mipsel and mipseb in all the places that we have just mips.  And there are other complications with doing that (binutils doesn't like to build mips tools without both byte genders, it seems).
Introduced BINUTIL_ARCH so that other bisexual architectures can a
generic mechanism.

We cannot just define MACHINE_ARCH as mips because we need to
differentiate big and little endian types of binaries.  Discussions on
freebsd-arch have hashed out this issue (and the parallel libc
issues).  NetBSD is moving towards mipsel and mipseb for their two
flavors of mips ports (in time for 1.4, if this change hasn't already
been accomplished).

I've been building i386 worlds with this tree for a three months with
these files in place with no ill effects.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@44360 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f