Close PR #10264. Don't bail directly out of passwd/chpass in my_yp_match(). Instead, treat the inability to retrieve a record from the server as a match failure and let things take its course.
Part of the problem here is that NIS _is_ turned on, however the master
server is actually not an NIS server: it's an NIS+ server. And the client
is bound to an NIS+ replica server that's running in YP compat mode.
The code which tries to figure out of the user is local or NIS gets
confused by this.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@44375 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f