o Migrate path, umask from dot.login to dot.cshrc o Comment out display of fortune by default. o Synch root's .cshrc/.login and non-root's .cshrc/.login in terms of   gratuitous variables set (EDITOR). o Remove some commented out variables set inconsistently or gratuitously,   such as Interviews settings, 8-bit German locale for root only. o Synchronize comments in header, as well as references to appropriate man   pages. o Remove MANPATH setting as apparently /etc/manpath.config does all that   already.
Similar changes probably need to be made in other dot.* files for root
and skel, as all of these files seem to set different aliases, environmental
variables, prompts, and have different semantics.

As a result of this patch, leaving aside the setting of a special prompt
for root, users of csh and tcsh should find similar environments when
logging in or su'ing to any account using that shell.

Reviewed by:	asmodai, nbm, will


git-svn-id: svn+ssh://svn.freebsd.org/base/head@58570 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f